misc: fpc1020: Set fingerprintd priority to max when screen is off
Modified for the Zuk Z2 by OrdenKrieger. Of course we have another fp driver for the fpc1020 then every other device with the same fp sensor.
Give fingerprintd the highest priority (MIN_NICE) when the screen is off to
speed up fingerprint processing, and then reset its priority back to normal
when the screen is on.

Signed-off-by: Sultanxda <sultanxda@gmail.com>