treewide: Fix incorrect debugfs init error checks
When CONFIG_DEBUGFS is disabled, all of the debugfs functions that return
a pointer are hardcoded to return ERR_PTR(-ENODEV). This causes debugfs
error checks in many drivers to not work correctly because they only check
to see if the return value is NULL, rather than checking to see if it is
NULL or contains an error code.

This fixes the incorrect debugfs error check assumptions, as well as fixes
init routines where debugfs failure was treated as a fatal error (so that
the drivers will still work when CONFIG_DEBUGFS is disabled).

Signed-off-by: Sultanxda <sultanxda@gmail.com>