sched/fair: cleanup select_energy_cpu_brute to be more consistent
The current definition of select_energy_cpu_brute is a bit confusing in
the definition of the value for the target_cpu to be returned as wakeup
CPU for the specified task.

This cleanup the code by ensuring that we always set target_cpu right
before returning it. rcu_read_lock and check on *sd!=NULL are also moved
around to be exactly where they are required.

Change-Id: I70a4b558b3624a13395da1a87ddc0776fd1d6641
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: Chris Redpath <chris.redpath@arm.com>
Signed-off-by: Quentin Perret <quentin.perret@arm.com>
[@nathanchance: Backport to 4.4, adjust context]
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
[@yarost12: backport to 3.18, removed 'eenv.trg_cpu = backup_cpu;'
causes bootloops in 3.18]
Signed-off-by: YaroST12 <yaro330@gmail.com>