sched/fair: add SchedTune's PESpace filtering
Currently, the best schedule candidate CPU is selected by looking just
at the difference in energy consumption between two candidate. A
dead-zone filter is also applied to avoid task wakeup migrations when
energy is not significantly impacted.

This patch adds support for another filtering policy which is based on
SchedTune's Performance/Energy evaluation. A sched_feature allows to
fallback to the original eas_deadzone_filter policy, which is also used
for non boosted tasks. For boosted tasks instead this patch exploits the
SchedTune provided functionalities to evaluate the performance-vs-energy
tradeoff.

Change-Id: Iee83f96efe8693ab4f3fc67c6d06bb0550405fec
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: Chris Redpath <chris.redpath@arm.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>