sched/tune: Move CPU and task boost margin APIs into tune.c
Below three functions are related with boost margin calculations:

  schedtune_margin()
  schedtune_cpu_margin()
  schedtune_task_margin

These three functions belong to schedtune and they are called in fair.c
to calculate the boosted utilization for CPU and task; in order to let
these three functions to be more friendly for callers in fair.c, this
patch is to move functions definition into tune.c and add declaration
in tune.h.

Current code defines the different return value type for these three
functions, some of them are 'int' type and some are 'long' type. This
patch also polishes return value type for 'long' type.

Change-Id: Icfe89efbeebabf28cc008b69306d36088069489a
Signed-off-by: Leo Yan <leo.yan@linaro.org>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>