misc: cclogic: wake screen up when cable is attached/detached
Cclogic driver sometimes fails to wake system up properly,
this results in userspace not being able to start charging
or properly connect MTP.

Fix this by adding a "backup" plan for such cases, it will be
delayed by 30ms, which is enough for cclogic to wake system
up when it's not deep sleeping, but not too intrusive to
cause issues on other events.

Only allow 1 execution on last cclogic_irq event
(there are often 3-4 on each attach/detach) by adding
a cancel right before scheduling and cancel when the detach
event occurs second time (for whatever reason).

Signed-off-by: Yaroslav Furman <yaro330@gmail.com>